Separate campaign header to it's own component
